Animal Feeding Game
I developed this project as part of the Unity Junior Programming Course. The assets and other content belong to the Unity course. In this project, we see the screen from a vertical angle. We feed animals running towards our character and prevent them from passing behind us. Additionally, I coded a different variation of the Tetris game.